--- a/buildenv/jenkins/docker-slaves/ppc64le/centos7/Dockerfile
+++ b/buildenv/jenkins/docker-slaves/ppc64le/centos7/Dockerfile
@@ -110,17 +110,23 @@ RUN cd /usr/local \
   && rm -rf gcc-7.tar.xz \
   && ln -s /usr/local/bin/gcc-7.3 /usr/bin/cc \
   && ln -s /usr/local/bin/gcc-7.3 /usr/bin/gcc \
-  && ln -s /usr/local/bin/g++-7.3 /usr/bin/g++ 
+  && ln -s /usr/local/bin/g++-7.3 /usr/bin/g++
+
+RUN cd /tmp \
+  && wget https://github.com/protocolbuffers/protobuf/releases/download/v3.7.1/protobuf-cpp-3.7.1.tar.gz \
+  && tar -xzf protobuf-cpp-3.7.1.tar.gz \
+  && rm -f protobuf-cpp-3.7.1.tar.gz \
+  && cd ./protobuf-3.7.1 \
+  && ./configure --disable-shared --with-pic \
+  && make \
+  && make install \
+  && cd .. \
+  && rm -rf /tmp/protobuf-3.7.1
 
 # Edit ldconfig to connect this library
 RUN echo "/usr/local/lib64" >> /etc/ld.so.conf.d/local.conf \
+  echo "/usr/local/lib" >> /etc/ld.so.conf.d/local.conf \
   && ldconfig
-  
-# Dependency required by test framework
-RUN wget -O - http://cpanmin.us | perl - --self-upgrade \
-  && cpanm Text::CSV \
-  && cpanm JSON \
-  && cpanm XML::Parser
 
 # Install cmake version 3.11.4
 RUN cd /tmp \
@@ -131,7 +137,7 @@ RUN cd /tmp \
   && ./configure \
   && make \
   && make install
-  
+
 #Building and setting up git version 2.5.3
 RUN cd /usr/src \
   && wget https://www.kernel.org/pub/software/scm/git/git-2.5.3.tar.gz \