Build against OMR `b372b24c8cb05c9a26f3fc21ffe4e8817ae488d6`: eclipse/openj9-omr#44