@a7ehuo 2019-11-14
b9c7c3d updated the files to eliminate the difference from the `master` branch on extra lines. 

With b9c7c3d and after this rebase, there are still files that are different between the `master` and the `jitaas` branch in three categories: 

1. Files that still need to be updated/merged into the `master` branch from the `jitaas` branch 
```
LICENSE <= https://github.com/eclipse/openj9/pull/6091
longabout.html <=  https://github.com/eclipse/openj9/pull/6091

runtime/compiler/build/toolcfg/gnu/common.mk <= https://github.com/eclipse/openj9/pull/6990
runtime/compiler/codegen/J9TreeEvaluator.cpp <= https://github.com/eclipse/openj9/pull/6909

runtime/compiler/env/VMJ9.cpp <= https://github.com/eclipse/openj9/pull/6909
runtime/compiler/env/VMJ9.h <= https://github.com/eclipse/openj9/pull/6909
runtime/compiler/env/VMJ9Server.cpp <= https://github.com/eclipse/openj9/pull/6909
runtime/compiler/env/VMJ9Server.hpp <= https://github.com/eclipse/openj9/pull/6909
runtime/compiler/il/J9Node.cpp <= https://github.com/eclipse/openj9/pull/7391
runtime/compiler/ilgen/Walker.cpp <= https://github.com/eclipse/openj9/pull/7385
runtime/compiler/net/protos/compile.proto <= https://github.com/eclipse/openj9/pull/6909
runtime/compiler/p/codegen/CallSnippet.cpp <= https://github.com/eclipse/openj9/pull/7386
runtime/compiler/p/codegen/J9PPCInstruction.cpp <= https://github.com/eclipse/openj9/pull/7384
runtime/compiler/p/codegen/PPCJNILinkage.cpp <=https://github.com/eclipse/openj9/pull/7391
runtime/compiler/p/runtime/Recomp.cpp <= https://github.com/eclipse/openj9/pull/7392
runtime/compiler/z/codegen/J9TreeEvaluator.cpp <= https://github.com/eclipse/openj9/pull/6909

runtime/compiler/control/CompilationController.hpp <= update copyright to 2019
runtime/compiler/control/rossa.cpp <= remove white trailing 
runtime/compiler/runtime/CMakeLists.txt <= update to properly indent with tab
test/docs/OpenJ9TestUserGuide.md <= Update JITServer test info 
```

2. Files that will or need to be updated on the `jitaas` branch
```
runtime/compiler/control/JITClientCompilationThread.cpp <= https://github.com/eclipse/openj9/pull/7731
runtime/compiler/control/JITServerCompilationThread.cpp <= https://github.com/eclipse/openj9/pull/7731
runtime/compiler/env/j9method.cpp <= https://github.com/eclipse/openj9/pull/7731
runtime/compiler/env/j9methodServer.cpp <= https://github.com/eclipse/openj9/pull/7731

runtime/compiler/env/JITServerPersistentCHTable.cpp <= TODO: TR_ASSERT or TR_ASSERT_FATAL
```

3. Files that should remain different between the `master` branch and the `jitaas` branch.

```
runtime/compiler/CMakeLists.txt
runtime/compiler/compiler.mk

buildenv/docker/jdk8/s390x/ubuntu16/jitserver/build/Dockerfile
buildenv/docker/jdk8/s390x/ubuntu16/jitserver/buildenv/Dockerfile
buildenv/docker/jdk8/x86_64/rhel7/jitserver/build/Dockerfile
buildenv/docker/jdk8/x86_64/ubuntu16/jitserver/build/Dockerfile
buildenv/docker/jdk8/x86_64/ubuntu16/jitserver/buildenv/Dockerfile
buildenv/docker/jdk8/x86_64/ubuntu18/jitserver/build/Dockerfile
buildenv/docker/jdk8/x86_64/ubuntu18/jitserver/buildenv/Dockerfile
doc/compiler/jitserver/Build.md
doc/compiler/jitserver/Docker.md
doc/compiler/jitserver/Usage.md
```