@a7ehuo 2019-11-19
Looked at the rebased files in this PR under  `runtime/compiler/env/` and `runtime/compiler/runtime`, so far looks like only #7724 needs to be implemented for `JITServer`.