@a7ehuo 2019-11-29
@mpirvu Ready for review. Thanks! I ran acmeair and didn't observe performance drop. The results are as below: 

=== With master rebase ===
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-rebased/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:cacheDir=/opt/IBM/DT7/scripts/ahuoScripts/AcmeAir-mpirvu/AcmeAir/.classCache,name=libertyscc -Xmx256m -Xscmx200m -Xscmaxaot150m -Xaot:forceAOT
Throughput	avg=11312.64	min=11075.90	max=11601.00	stdDev=145.1	maxVar=4.74%	confInt=0.74%	samples=10
Intermediate results:
Run 0	981.4	10968.8	11075.9	Avg=11076	CPU=18387 ms  Footprint=258660 KB
Run 1	1339.1	11176.0	11281.2	Avg=11281	CPU=7473 ms  Footprint=253228 KB
Run 2	1385.0	11337.4	11353.0	Avg=11353	CPU=7451 ms  Footprint=255636 KB
Run 3	1369.3	11407.0	11388.3	Avg=11388	CPU=7433 ms  Footprint=253768 KB
Run 4	1353.6	10939.1	11180.7	Avg=11181	CPU=7135 ms  Footprint=260072 KB
Run 5	1404.7	11430.6	11601.0	Avg=11601	CPU=7624 ms  Footprint=264620 KB
Run 6	1400.2	11331.8	11435.0	Avg=11435	CPU=6957 ms  Footprint=261268 KB
Run 7	1394.3	11151.9	11290.0	Avg=11290	CPU=7396 ms  Footprint=261220 KB
Run 8	1389.9	11127.0	11303.8	Avg=11304	CPU=7390 ms  Footprint=261252 KB
Run 9	1379.2	11188.6	11217.5	Avg=11218	CPU=7358 ms  Footprint=251004 KB
CompTime	avg=8460.40	min=6957.00	max=18387.00	stdDev=3492.8	maxVar=164.29%	confInt=23.93%	samples=10
Footprint	avg=258072.80	min=251004.00	max=264620.00	stdDev=4413.8	maxVar=5.42%	confInt=0.99%	samples=10
```

==== without master rebase === 
```
Results for JDK=/opt/IBM/DT7/scripts/ahuoScripts/sdk/j2sdk-image-jitaas/j2sdk-image jvmOpts=-XX:+UseJITServer -Xjit:dontDowngradeToCold,sampleThreshold=0,scorchingSampleThreshold=0 -Xshareclasses:cacheDir=/opt/IBM/DT7/scripts/ahuoScripts/AcmeAir-mpirvu/AcmeAir/.classCache,name=libertyscc -Xmx256m -Xscmx200m -Xscmaxaot150m -Xaot:forceAOT
Throughput	avg=11291.94	min=10917.60	max=11482.60	stdDev=184.1	maxVar=5.18%	confInt=0.95%	samples=10
Intermediate results:
Run 0	982.9	10823.2	10917.6	Avg=10918	CPU=24824 ms  Footprint=247464 KB
Run 1	1310.3	11045.7	11094.6	Avg=11095	CPU=8290 ms  Footprint=265640 KB
Run 2	1336.3	11248.3	11346.4	Avg=11346	CPU=8310 ms  Footprint=264984 KB
Run 3	1339.2	11148.4	11247.5	Avg=11248	CPU=8344 ms  Footprint=264684 KB
Run 4	1312.7	10993.1	11162.5	Avg=11162	CPU=7495 ms  Footprint=257900 KB
Run 5	1335.2	11255.0	11397.0	Avg=11397	CPU=7314 ms  Footprint=263520 KB
Run 6	1349.5	11442.2	11462.0	Avg=11462	CPU=7865 ms  Footprint=269504 KB
Run 7	1345.6	11233.1	11440.7	Avg=11441	CPU=7657 ms  Footprint=260280 KB
Run 8	1347.0	11207.8	11368.5	Avg=11368	CPU=7455 ms  Footprint=261976 KB
Run 9	1342.7	11457.9	11482.6	Avg=11483	CPU=7628 ms  Footprint=262736 KB
CompTime	avg=9518.20	min=7314.00	max=24824.00	stdDev=5391.3	maxVar=239.40%	confInt=32.83%	samples=10
Footprint	avg=261868.80	min=247464.00	max=269504.00	stdDev=5957.4	maxVar=8.91%	confInt=1.32%	samples=10
```